// UniverCity-ui/src/app/Services/models/page-response-university-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UniversityResponse } from '../models/university-response';
export interface PageResponseUniversityResponse {
  content?: Array<UniversityResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}